refactor: extract note utility functions into separate module for better code organization
